% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navervoice.R
\name{naverVoice}
\alias{naverVoice}
\title{naverVoice}
\usage{
naverVoice(text = "", client_Id = "", client_secret = "",
  speaker = c("mijin", "jinho", "clara", "matt", "yuri", "shinji", "meimei"),
  speed = 0)
}
\arguments{
\item{client_Id}{Naver Open API Client ID}

\item{client_secret}{Naver Open API Client secret}

\item{speaker}{speaker whose voice is composed'
refer to below
majin : Korean woman, jinhi : Korean Man, clara : American woman, matt : American man, yuri : Japanses woman, shinji : Japanese man, meimei : Chinese woman}
}
\description{
Make a voice synthesis using Naver Open API
}
\details{
It helps you to make a various voice synthesis using Naver Open API
}
\examples{
naverVoice('best','DpMY9iw4AztbeSxYXf2t','69Az0lv56T')
}
\author{
Kim BM(ksd2080@hotmail.com)
}

