% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blog_search.R
\docType{package}
\name{naverblog.search}
\alias{naver.search}
\alias{naverblog.search}
\alias{naverblog.search-package}
\title{naverblog.search}
\usage{
naver.search(keyword = "", client_Id = "", client_secret = "",
  category = c("blog", "news", "region", "encyclopedia", "movie"),
  display = 100, start = 1, sort = "sim", genre = "", country = "",
  yearfrom = NULL, yearto = NULL)
}
\arguments{
\item{keyword}{keyword you want to search}

\item{client_Id}{Naver Open API Client ID}

\item{client_secret}{Naver Open API Client secret}

\item{category}{category whcich you want to search (e.g : blog, news, region,.....)}

\item{display}{the number of search result}

\item{start}{start number of search result}

\item{sort}{sim : similarity, date : ordered by date}
}
\description{
Searching Naver blog with using Naver Open API
}
\details{
This is a package which use Naver Open API to search naver blog You must need Naver client ID and secret
}
\examples{
naverblog.search(keyword = 'andong', client_Id = 'DpMY9iw4AztbeSxYXf2t', client_secret = '69Az0lv56T')

}
\author{
Kim BM(ksd2080@hotmail.com)
}

